+++ shardingsphere-operator/pkg/kubernetes/container/container.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package common
+package container
 
 import (
 	v1 "k8s.io/api/core/v1"
