+++ shardingsphere-operator/pkg/kubernetes/container/container_test.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package common
+package container
 
 import (
 	"testing"
